+++ packages/core/src/core.ts
@@ -3,8 +3,6 @@ import { ethers } from 'ethers';
 // Our core library
 export * from './certificate';
 export * from './clause';
-export * from './encoding';
-export * from './vcdm/hash';
 export * from './hdkey';
 export * from './keystore';
 export * from './secp256k1';
