+++ packages/core/src/utils/const/transaction.ts
@@ -1,4 +1,4 @@
-import { RLP_CODER } from '../../encoding';
+import { RLP_CODER } from '../../vcdm';
 
 /**
  * Transaction gas constants
@@ -152,11 +152,11 @@ const SIGNATURE_LENGTH = 65;
 const BLOCK_REF_LENGTH = 8;
 
 export {
-    TRANSACTIONS_GAS_CONSTANTS,
-    UNSIGNED_TRANSACTION_RLP,
+    BLOCK_REF_LENGTH,
+    SIGNATURE_LENGTH,
     SIGNED_TRANSACTION_RLP,
     TRANSACTION_FEATURES_KIND,
     TRANSACTION_SIGNATURE_KIND,
-    SIGNATURE_LENGTH,
-    BLOCK_REF_LENGTH
+    TRANSACTIONS_GAS_CONSTANTS,
+    UNSIGNED_TRANSACTION_RLP
 };
