+++ packages/core/src/vcdm/encoding/index.ts
