+++ packages/core/src/vcdm/encoding/rlp/helpers/fixedhexblobkind.ts
