+++ packages/core/src/vcdm/encoding/rlp/helpers/hexblobkind.ts
@@ -1,4 +1,4 @@
-import { Hex } from '../../../vcdm/Hex';
+import { Hex } from '../../../../vcdm';
 import { InvalidRLP } from '@vechain/sdk-errors';
 import { type RLPInput } from '../types';
