+++ packages/core/src/vcdm/encoding/rlp/helpers/index.ts
