+++ packages/core/src/vcdm/encoding/rlp/kind/bufferkind.ts
