+++ packages/core/src/vcdm/encoding/rlp/kind/hexblobkinds/fixedhexblob.ts
