+++ packages/core/src/vcdm/encoding/rlp/kind/hexblobkinds/index.ts
