+++ packages/core/src/vcdm/encoding/rlp/kind/hexblobkinds/optionalfixedhexblob.ts
