+++ packages/core/src/vcdm/encoding/rlp/kind/numerickind.ts
