+++ packages/core/src/vcdm/encoding/rlp/kind/scalarkind.abstract.ts
