+++ packages/core/src/vcdm/encoding/rlp/rlp.ts
