+++ packages/core/src/vcdm/encoding/rlp/types.d.ts
