+++ packages/core/tests/vcdm/encoding/rlp/helpers.fixture.ts
