+++ packages/core/tests/vcdm/encoding/rlp/helpers.unit.test.ts
@@ -1,6 +1,10 @@
 import { describe, expect, test } from '@jest/globals';
 import {
-    Hex,
+    InvalidDataType,
+    InvalidRLP,
+    stringifyData
+} from '@vechain/sdk-errors';
+import {
     assertCompactFixedHexBlobBuffer,
     assertFixedHexBlobKindBuffer,
     assertFixedHexBlobKindData,
@@ -9,8 +13,9 @@ import {
     assertValidNumericKindBuffer,
     decodeBufferToHexWithLeadingZeros,
     encodeBigIntToBuffer,
+    Hex,
     validateNumericKindData
-} from '../../src';
+} from '../../../../src';
 import {
     invalidCompactFixedHexBlobKindBufferTestCases,
     invalidFixedHexBlobKindBufferTestCases,
@@ -27,11 +32,6 @@ import {
     validHexBlobKindDataTestCases,
     validNumericBufferTestCases
 } from './helpers.fixture';
-import {
-    InvalidDataType,
-    InvalidRLP,
-    stringifyData
-} from '@vechain/sdk-errors';
 
 /**
  * Test suite for BigInt helper functions
