+++ packages/core/tests/vcdm/encoding/rlp/rlp.fixture.ts
@@ -1,4 +1,8 @@
-import { RLP_CODER, type RLPProfile, type RLPValidObject } from '../../src';
+import {
+    RLP_CODER,
+    type RLPProfile,
+    type RLPValidObject
+} from '../../../../src';
 
 /* Simple RLP_CODER encode */
 const encodeTestCases = [
@@ -853,49 +857,49 @@ const decodeMixedKindProfileTestCases = [
 ];
 
 export {
-    encodeTestCases,
-    decodeTestCases,
-    numericKindEncodeTestCases,
-    numericKindDecodeTestCases,
-    invalidNumericKindEncodeTestCases,
-    invalidNumericKindDecodeTestCases,
-    invalidBufferKindDecodeTestCases,
-    bufferProfile,
     bufferData,
+    bufferProfile,
+    compactFixedHexBlobKindDecodeTestCases,
+    compactFixedHexBlobKindEncodeTestCases,
+    decodeBufferProfileTestCases,
+    decodeCompactFixedHexBlobProfileTestCases,
+    decodeHexBlobProfileTestCases,
+    decodeMixedKindProfileTestCases,
+    decodeNumericProfileTestCases,
+    decodeTestCases,
+    encodeBufferProfileTestCases,
+    encodeCompactFixedHexBlobProfileTestCases,
+    encodeFixedHexBlobProfileTestCases,
+    encodeHexBlobProfileTestCases,
+    encodeMixedKindProfileTestCases,
+    encodeNumericProfileTestCases,
+    encodeOptionalFixedHexBlobProfileTestCases,
+    encodeTestCases,
+    fixedHexBlobData,
+    fixedHexBlobKindDecodeTestCases,
+    fixedHexBlobKindEncodeTestCases,
+    fixedHexBlobProfile,
+    hexBlobData,
+    hexBlobKindDecodeTestCases,
+    hexBlobKindEncodeTestCases,
+    hexBlobProfile,
     invalidBufferData,
-    numericProfile,
+    invalidBufferKindDecodeTestCases,
+    invalidDecodeObjectTestCases,
+    invalidEncodeObjectTestCases,
+    invalidFixedBlobKindDecodeTestCases,
+    invalidFixedHexBlobEncodeTestCases,
+    invalidHexBlobKindDecodeTestCases,
+    invalidHexBlobKindEncodeTestCases,
+    invalidNumericKindDecodeTestCases,
+    invalidNumericKindEncodeTestCases,
     numericData,
-    numericProfileWithMaxBytes,
-    numericDataWithMaxBytes,
-    numericDataWithString,
     numericDataInvalidArray,
     numericDataWithByteOverflow,
-    encodeNumericProfileTestCases,
-    encodeBufferProfileTestCases,
-    invalidEncodeObjectTestCases,
-    decodeBufferProfileTestCases,
-    invalidDecodeObjectTestCases,
-    hexBlobKindEncodeTestCases,
-    decodeNumericProfileTestCases,
-    decodeHexBlobProfileTestCases,
-    hexBlobProfile,
-    hexBlobData,
-    encodeHexBlobProfileTestCases,
-    invalidHexBlobKindEncodeTestCases,
-    hexBlobKindDecodeTestCases,
-    invalidHexBlobKindDecodeTestCases,
-    fixedHexBlobProfile,
-    fixedHexBlobData,
-    encodeFixedHexBlobProfileTestCases,
-    fixedHexBlobKindEncodeTestCases,
-    invalidFixedHexBlobEncodeTestCases,
-    fixedHexBlobKindDecodeTestCases,
-    invalidFixedBlobKindDecodeTestCases,
-    encodeOptionalFixedHexBlobProfileTestCases,
-    encodeCompactFixedHexBlobProfileTestCases,
-    decodeCompactFixedHexBlobProfileTestCases,
-    encodeMixedKindProfileTestCases,
-    decodeMixedKindProfileTestCases,
-    compactFixedHexBlobKindEncodeTestCases,
-    compactFixedHexBlobKindDecodeTestCases
+    numericDataWithMaxBytes,
+    numericDataWithString,
+    numericKindDecodeTestCases,
+    numericKindEncodeTestCases,
+    numericProfile,
+    numericProfileWithMaxBytes
 };
